Pixmaps with size less than 100px doesn't crashes sankore on mouse click.

+++ src/domain/UBGraphicsPixmapItem.cpp
@@ -65,8 +65,9 @@ void UBGraphicsPixmapItem::mousePressEvent(QGraphicsSceneMouseEvent *event)
     QMimeData* pMime = new QMimeData();
     pMime->setImageData(pixmap().toImage());
     mDelegate->setMimeData(pMime);
-    int k = pixmap().width() / 100;
-    QSize newSize(pixmap().width() / k, pixmap().height() / k);
+    qreal k = (qreal)pixmap().width() / 100.0;
+
+    QSize newSize((int)(pixmap().width() / k), (int)(pixmap().height() / k));
 
     mDelegate->setDragPixmap(pixmap().scaled(newSize, Qt::IgnoreAspectRatio, Qt::SmoothTransformation));
